requirements 2.txt
# Core dependencies
langgraph>=0.2.0
langchain-core>=0.3.0
typing-extensions>=4.0.0
openai>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0

# Database
supabase>=2.0.0

# Geospatial
geopy>=2.3.0
haversine>=2.8.0

# Routing and optimization
networkx>=3.0
scipy>=1.10.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
cachetools>=5.3.0

# Async support
aiohttp>=3.8.0